♻️[Refactor] 모임 개설 상태 관리 유틸 함수로 분리 #204
